requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
pdfkit==1.0.0
unidecode==1.3.7
mysql-connector-python==8.2.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pymysql==1.1.0
schedule==1.2.0
pytz==2023.3 